% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg class}
\description{
This class contains methods to handle multiple linear regression.
You can find more information about the linear regression in the link
bellow:
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula to show the relationship between
dependent and independent variables.}

\item{\code{data}}{A data.frame to give the values of variables.}

\item{\code{reg_coeffs}}{A matrix of regression coefficients.}

\item{\code{fitted_value}}{A matrix of predicted values for dependent
variable.}

\item{\code{resids}}{A matrix that contains the differences between
predicted and actual values of dependent variable.}

\item{\code{df}}{A numeric scalar that shows the degree of freedom.}

\item{\code{resids_var}}{A matrix that shows the variance of residuals.}

\item{\code{reg_coef_var}}{A matrix that shows the variance of regression
coefficients.}

\item{\code{t_values}}{A matrix that shows t values.}

\item{\code{p_values}}{A matrix of p values.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{printing regression coefficients}

\item{\code{initialize(formula, data)}}{initializing regression calculations}

\item{\code{plot()}}{ploting residuals and scaled residuals vs fitted values}

\item{\code{pred()}}{printing vector of fitted values}

\item{\code{print()}}{print regression coefficients}

\item{\code{resid()}}{printing vector of resisuals}
}}

